classification/requirements.txt
torch==2.2.2
torchvision==0.17.2
timm==0.9.16
opencv-python==4.10.0.84
numpy==1.26.4
tqdm==4.66.4
matplotlib
scikit-learn
Pillow
rad-dino==0.1.1
wandb
transformers
einops
jaxtyping